==> 1-data wrangling.ipynb <==
import numpy as np
import pandas as pd
import warnings

pd.set_option('display.max_columns', None)
warnings.filterwarnings('ignore')

high popularity songs dataset from spotify

data_high_popularity = pd.read_csv("datasets/high_popularity_spotify_data.csv")
data_high_popularity.head()

low popularity songs dataset from spotify

data_low_popularity = pd.read_csv("datasets/low_popularity_spotify_data.csv")
data_low_popularity.head()

all column names in high popularity dataset

data_high_popularity.columns.sort_values()

all column names in low popularity dataset

data_low_popularity.columns.sort_values()

both datasets have same set of columns

missing values in high popularity dataset

data_high_popularity.isnull().sum()

data_high_popularity[data_high_popularity['track_album_name'].isnull()]

the album name of album id "5zQI9dFbS9TrhvC9clgjz7" is "NA"

data_high_popularity.loc[665, 'track_album_name'] = "NA (EP)"

data_high_popularity.isnull().sum()

now there are no missing values in the high popularity dataset

missing values in low popularity dataset

data_low_popularity.isnull().sum()

it is not possible to get music specific data such as loudness and tempo, so better drop all null values from this dataset

data_low_popularity.dropna(inplace=True)

data_low_popularity.isnull().sum()

all missing values are cleaned from low popularity dataset

now combine both datasets into a single dataset

set a column "popularity" which is 1 for high popular songs and 0 for low popular songs

data_high_popularity['popularity'] = 1
data_low_popularity['popularity'] = 0
data = pd.concat([data_high_popularity, data_low_popularity], ignore_index=True)

data.head()

data.tail()

data.isnull().sum()

first look at audio features

audio_features = ['energy', 'tempo', 'danceability', 'loudness', 'liveness', 
                  'valence', 'speechiness', 'instrumentalness', 'mode', 'key', 
                  'duration_ms', 'acousticness']

data[audio_features].describe().T

for feature in audio_features:
    print(f'Length of unique values of {feature}: {len(data[feature].unique())}')

"mode" and "key" are categorical features

data['key'].value_counts()

now the descriptive features

data['playlist_genre'].unique()

data['time_signature'].unique()

both "playlist-genre" and "time_signature" are categorical features

len(data['playlist_subgenre'].unique())

len(data['playlist_name'].unique())

these features have a wide range of categorical data and are hard to analyse

len(data['type'].unique())

type for all the entries is "audio_features"

also "track_href", "uri", "analysis_url", "track_album_id", "track_id", "playlist_id" attributes are unique to each entry and can be removed from the dataset. [keeping the "id" feature as a unique identifer for each entry]

data.drop(columns=["playlist_subgenre", "playlist_name", "track_href", "track_id", "type", "playlist_id",
                   "uri", "analysis_url", "track_album_id"], inplace=True)

data.head()

get the year of release of each track

data['release_date'] = pd.to_datetime(data['track_album_release_date'], errors='coerce')
data.head()

data['release_date'].isnull().sum()

data.loc[data['release_date'].isnull(), 'track_album_release_date'].unique()

data['release_year'] = data['release_date'].dt.year

data['release_year'].isnull().sum()

fill the null values of "release_year" with data from "track_album_release_date"

data.loc[data['release_year'].isnull(), 'release_year'] = data.loc[data['release_year'].isnull(), 'track_album_release_date'].str[:4]

data['release_year'].isnull().sum()

data.head()

data['release_year'] = data['release_year'].astype(int)

data.drop(columns=['track_album_release_date', 'release_date'], inplace=True)

data.rename(columns={'track_artist':'artist', 'track_album_name':'album',
             'track_name':'name', 'playlist_genre':'genre'}, inplace=True)
data.head()

data.columns[data.isnull().any()]

the data set has no null values now

data.duplicated().sum()

drop the duplicates

data = data.drop_duplicates()
data.duplicated().sum()

the dataset is cleaned and missing values are replaced 

data.to_csv("datasets/data.csv", index=False)
