==> 3-exploratory data analysis (sql).ipynb <==
%load_ext sql

import csv, sqlite3
import pandas as pd

con = sqlite3.connect("database.db")
cur = con.cursor()

%sql sqlite:///database.db

data = pd.read_csv('datasets/data.csv')
data.to_sql("SONGDATA", con, if_exists='replace', index=False, method="multi", chunksize=1000)

description of the table

%sql PRAGMA table_info(SONGDATA)

<h3>data for popular songs in 2024</h3>

lets see top 10 most popular songs in 2024

%%sql
SELECT DISTINCT (name), artist FROM SONGDATA 
WHERE release_year = 2024
ORDER BY track_popularity DESC LIMIT 10

as pop is the most popular genre, lets see how many pop songs were popular in 2024

%%sql
SELECT COUNT(*) AS POPULAR_SONGS FROM SONGDATA 
WHERE release_year = 2024 AND genre = "pop" AND popularity = 1

there are 62 pop songs released in 2024 that were popular

now lets see the songs that were really popular

%%sql
SELECT COUNT(*) AS VERY_POPULAR_SONGS FROM SONGDATA 
WHERE release_year = 2024 AND genre = "pop" AND track_popularity > 90

there were 12 very popular songs, lets see what they were

%%sql
SELECT name, artist, track_popularity FROM SONGDATA 
WHERE release_year = 2024 AND genre = "pop" AND track_popularity > 90
GROUP BY name
ORDER BY track_popularity DESC

<h4>songs and popularity of few famous artists</h4>

Bruno Mars

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Bruno Mars%"
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Sabrina Carpenter

%%sql
SELECT DISTINCT name, artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Sabrina Carpenter%"
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Taylor Swift

%%sql
SELECT DISTINCT name, artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Taylor Swift%"
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Drake

%%sql
SELECT DISTINCT name, artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Drake%"
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Billie Eilish

%%sql
SELECT DISTINCT name, artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Billie Eilish%"
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

all of these artists had few popular songs in 2024

<h4>2nd popular song genres in 2023 and 2024</h4>

most popular k-pop genre songs

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE genre = 'k-pop'
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

lets see the most popular k-pop genre song in 2023

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE genre = 'k-pop' AND release_year = 2023
GROUP BY name
ORDER BY track_popularity DESC LIMIT 1

most popular gaming genre songs

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE genre = 'gaming'
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

lets see most popular gaming genre song in 2024

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE genre = 'gaming' AND release_year = 2024
GROUP BY name
ORDER BY track_popularity DESC LIMIT 1

<h3> 2024 popular songs of famous artists</h3>

Bruno Mars

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Bruno Mars%" AND release_year = 2024
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Sabrina Carpenter

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Sabrina Carpenter%" AND release_year = 2024
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Taylor Swift

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Taylor Swift%" AND release_year = 2024
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Drake

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Drake%" AND release_year = 2024
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

Billie Eilish

%%sql
SELECT DISTINCT(name), artist, genre, release_year, track_popularity FROM SONGDATA
WHERE artist LIKE "%Billie Eilish%" AND release_year = 2024
GROUP BY name
ORDER BY track_popularity DESC LIMIT 5

<h4>How much popularity did songs of few famous artists gain in 2024?</h4>
1. 2 songs by Bruno Mars were very popular and topped the popularity list<br>
2. Sabrina Carpenter and Billie Eilish had few songs which were also very popular<br>
3. Songs released by Taylor Swift and Drake couldn't gain much popularity
