==> 4-feature engineering (for prediction).ipynb <==
import numpy as np
import pandas as pd
import warnings

pd.set_option('display.max_columns', None)
warnings.filterwarnings('ignore')

data = pd.read_csv("datasets/data.csv")
data.head()

data.duplicated().sum()

song specific data not needed for predicting whether a song will be popular or not

new_data = data.drop(columns=['id', 'artist', 'album', 'name', 'track_popularity', 'genre'])
new_data.head()

lets see how many records of songs are present which are released on and after 2000

new_data[new_data['release_year'] >= 2000].shape

new_data[new_data['release_year'] < 2000].shape

majority of the songs are after 2000, so lets take the songs which are recent

new_data = new_data[new_data['release_year'] >= 2000]

age of the songs calculated from 2025

new_data['age'] = 2025 - new_data['release_year']
new_data = new_data.drop(columns=['release_year'])

new_data.head()

new_data.tail()

num_features = ["energy", "tempo", "danceability", "loudness", "liveness", "valence", "instrumentalness", "duration_ms", "acousticness", "age", "speechiness"]

scaled_data = pd.DataFrame()
scaled_data['popularity'] = new_data['popularity']

scaling the numerical features

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()

for feature in num_features:
    scaled_data[feature] = scaler.fit_transform(new_data[[feature]])

hot encoding for categorical features

keydum = pd.get_dummies(new_data["key"], prefix="key").astype(int)
tsdum = pd.get_dummies(new_data["time_signature"], prefix="timesign").astype(int)

scaled_data = pd.concat([scaled_data, keydum, tsdum], axis=1)

scaled_data.tail(10)

save the data for prediction

scaled_data.to_csv("datasets/data_prediction.csv", index=False)
